+++ app/Http/Livewire/ContactPopup.php
@@ -3,6 +3,9 @@
 namespace App\Http\Livewire;
 
 use App\Mail\ContactMessageMail;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
 use Illuminate\Support\Facades\Mail;
 use Livewire\Component;
 use Lukeraymonddowning\Honey\Traits\WithHoney;
@@ -59,7 +62,14 @@ class ContactPopup extends Component
         $this->reset();
     }
 
-    public function render()
+    public function closePopUp()
+    {
+        $this->resetValidation();
+
+        $this->showContactMePopUp = false;
+    }
+
+    public function render(): Factory|View|Application
     {
         return view('livewire.contact-popup');
     }
