+++ app/Providers/ViewServiceProvider.php
@@ -2,8 +2,8 @@
 
 namespace App\Providers;
 
-use App\Http\ViewComposers\Fos\PostComposer;
-use App\Models\User;
+use App\Http\ViewComposers\Posts\CategoriesTagsViewComposer;
+use App\Http\ViewComposers\Posts\PostsListViewComposer;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
 
@@ -22,9 +22,8 @@ class ViewServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        View::composer(
-            ['fos.posts.create', 'fos.posts.edit'],
-            PostComposer::class
-        );
+        View::composers([
+            ['fos.posts.create', 'fos.posts.edit'], CategoriesTagsViewComposer::class,
+        ]);
     }
 }
